refactor: PR 반영
- 상수 변수 선언
- 토큰화 클래스 분리
