+++ src/main/kotlin/calcuator/StringAddCalculator.kt
@@ -2,9 +2,9 @@ package calcuator
 
 class StringAddCalculator {
     fun add(text: String?): Int {
-        if (text.isNullOrBlank()) return 0
+        if (text.isNullOrBlank()) return ZERO
 
-        val textOperands: List<String> = this.tokenizesText(text)
+        val textOperands: List<String> = TokenizeUtil.tokenizes(text)
 
         return textOperands
             .mapNotNull { textOperand -> textOperand.toIntOrNull() }
@@ -12,20 +12,7 @@ class StringAddCalculator {
             .sumOf { operand -> operand.number }
     }
 
-    private fun tokenizesText(text: String): List<String> {
-        val matcher = CUSTOM_DELIMITER.find(text)
-
-        if (matcher != null) {
-            val (customDelimiter: String, matchedText: String) = matcher.destructured
-
-            return matchedText.split(customDelimiter)
-        }
-
-        return text.split(DEFAULT_DELIMITER)
-    }
-
     companion object {
-        private val CUSTOM_DELIMITER = Regex("//(.)\n(.*)")
-        private val DEFAULT_DELIMITER = Regex("[,:]")
+        private const val ZERO = 0
     }
 }
